wren: Move cutsheet generation to own file
